drawer: extract reports tab
