+++ client/src/app/pages/applications/application-detail-drawer/application-detail-drawer.tsx
@@ -9,38 +9,20 @@ import {
   Tabs,
   Tab,
   TabTitleText,
-  Button,
-  DescriptionList,
-  DescriptionListDescription,
-  DescriptionListGroup,
-  DescriptionListTerm,
-  Divider,
-  Tooltip,
 } from "@patternfly/react-core";
 import spacing from "@patternfly/react-styles/css/utilities/Spacing/spacing";
-import CheckCircleIcon from "@patternfly/react-icons/dist/esm/icons/check-circle-icon";
-import ExclamationCircleIcon from "@patternfly/react-icons/dist/esm/icons/exclamation-circle-icon";
 
-import { Identity, MimeType, Manifest } from "@app/api/models";
-import { COLOR_HEX_VALUES_BY_NAME } from "@app/Constants";
-import { useFetchFacts } from "@app/queries/facts";
-import { useFetchIdentities } from "@app/queries/identities";
-import { useSetting } from "@app/queries/settings";
-import { getKindIdByRef } from "@app/utils/model-utils";
+import { Manifest } from "@app/api/models";
 
 import {
   IPageDrawerContentProps,
   PageDrawerContent,
 } from "@app/components/PageDrawerContext";
-import { EmptyTextMessage } from "@app/components/EmptyTextMessage";
 import { ReviewFields } from "@app/components/detail-drawer/review-fields";
 
-import DownloadButton from "./components/download-button";
-import { ApplicationFacts } from "./application-facts";
 import { formatPath } from "@app/utils/utils";
 import { Paths } from "@app/Paths";
 import { DecoratedApplication } from "../useDecoratedApplications";
-import { TaskStates } from "@app/queries/tasks";
 import { Toolbar, ToolbarContent, ToolbarItem } from "@patternfly/react-core";
 import { Table, Tbody, Td, Th, Thead, Tr } from "@patternfly/react-table";
 import { SimplePagination } from "@app/components/SimplePagination";
@@ -66,6 +48,7 @@ import { useFetchApplicationManifest } from "@app/queries/applications";
 import { usePlatformCoordinatesProvider } from "../usePlatformCoordinatesProvider";
 import { TabDetailsContent } from "./tab-details-content";
 import { TabTagsContent } from "./tab-tags-content";
+import { TabReportsContent } from "./tab-reports-contents";
 
 export interface IApplicationDetailDrawerProps
   extends Pick<IPageDrawerContentProps, "onCloseClick"> {
@@ -199,193 +182,6 @@ const TabPlatformCoordinatesContent: React.FC<{
       baseJsonDocument={document}
       jsonSchema={schema}
     />
-  );
-};
-
-const TabReportsContent: React.FC<{
-  application: DecoratedApplication;
-}> = ({ application }) => {
-  const { t } = useTranslation();
-  const { facts, isFetching } = useFetchFacts(application?.id);
-
-  const { identities } = useFetchIdentities();
-  let matchingSourceCredsRef: Identity | undefined;
-  let matchingMavenCredsRef: Identity | undefined;
-  if (application && identities) {
-    matchingSourceCredsRef = getKindIdByRef(identities, application, "source");
-    matchingMavenCredsRef = getKindIdByRef(identities, application, "maven");
-  }
-
-  const task = application.tasks.currentAnalyzer;
-  const taskState = task?.state ?? "";
-  const taskSucceeded = TaskStates.Success.includes(taskState);
-  const taskFailed = TaskStates.Failed.includes(taskState);
-
-  const notAvailable = <EmptyTextMessage message={t("terms.notAvailable")} />;
-
-  const enableDownloadSetting = useSetting("download.html.enabled");
-
-  const history = useHistory();
-  const navigateToAnalysisDetails = () =>
-    application?.id &&
-    task?.id &&
-    history.push(
-      formatPath(Paths.applicationsAnalysisDetails, {
-        applicationId: application?.id,
-        taskId: task?.id,
-      })
-    );
-
-  return (
-    <>
-      <TextContent className={spacing.mtMd}>
-        <Title headingLevel="h3" size="md">
-          Credentials
-        </Title>
-        {matchingSourceCredsRef && matchingMavenCredsRef ? (
-          <Text component="small">
-            <CheckCircleIcon color="green" />
-            <span className={spacing.mlSm}>Source and Maven</span>
-          </Text>
-        ) : matchingMavenCredsRef ? (
-          <Text component="small">
-            <CheckCircleIcon color="green" />
-            <span className={spacing.mlSm}>Maven</span>
-          </Text>
-        ) : matchingSourceCredsRef ? (
-          <Text component="small">
-            <CheckCircleIcon color="green" />
-            <span className={spacing.mlSm}>Source</span>
-          </Text>
-        ) : (
-          notAvailable
-        )}
-
-        <Title headingLevel="h3" size="md">
-          Analysis
-        </Title>
-        {taskSucceeded ? (
-          <>
-            <DescriptionList isHorizontal columnModifier={{ default: "2Col" }}>
-              <DescriptionListGroup>
-                <DescriptionListTerm>Details</DescriptionListTerm>
-                <DescriptionListDescription>
-                  <Tooltip content="View the analysis task details">
-                    <Button
-                      icon={
-                        <span className={spacing.mrXs}>
-                          <ExclamationCircleIcon
-                            color={COLOR_HEX_VALUES_BY_NAME.blue}
-                          ></ExclamationCircleIcon>
-                        </span>
-                      }
-                      type="button"
-                      variant="link"
-                      onClick={navigateToAnalysisDetails}
-                      className={spacing.ml_0}
-                      style={{ margin: "0", padding: "0" }}
-                    >
-                      View analysis details
-                    </Button>
-                  </Tooltip>
-                </DescriptionListDescription>
-
-                <DescriptionListTerm>Download</DescriptionListTerm>
-                <DescriptionListDescription>
-                  <Tooltip
-                    content={
-                      enableDownloadSetting.data
-                        ? "Click to download TAR file with HTML static analysis report"
-                        : "Download TAR file with HTML static analysis report is disabled by administrator"
-                    }
-                    position="top"
-                  >
-                    <DownloadButton
-                      application={application}
-                      mimeType={MimeType.TAR}
-                      isDownloadEnabled={enableDownloadSetting.data}
-                    >
-                      HTML
-                    </DownloadButton>
-                  </Tooltip>
-                  {" | "}
-                  <Tooltip
-                    content={
-                      enableDownloadSetting.data
-                        ? "Click to download YAML file with static analysis report"
-                        : "Download YAML file with static analysis report is disabled by administrator"
-                    }
-                    position="top"
-                  >
-                    <DownloadButton
-                      application={application}
-                      mimeType={MimeType.YAML}
-                      isDownloadEnabled={enableDownloadSetting.data}
-                    >
-                      YAML
-                    </DownloadButton>
-                  </Tooltip>
-                </DescriptionListDescription>
-              </DescriptionListGroup>
-            </DescriptionList>
-            <Divider className={spacing.mtMd}></Divider>
-          </>
-        ) : taskFailed ? (
-          task ? (
-            <>
-              <Button
-                icon={
-                  <span className={spacing.mrXs}>
-                    <ExclamationCircleIcon
-                      color={COLOR_HEX_VALUES_BY_NAME.red}
-                    ></ExclamationCircleIcon>
-                  </span>
-                }
-                type="button"
-                variant="link"
-                onClick={navigateToAnalysisDetails}
-                className={spacing.ml_0}
-                style={{ margin: "0", padding: "0" }}
-              >
-                Analysis details
-              </Button>
-            </>
-          ) : (
-            <span className={spacing.mlSm}>
-              <ExclamationCircleIcon
-                color={COLOR_HEX_VALUES_BY_NAME.red}
-              ></ExclamationCircleIcon>
-              Failed
-            </span>
-          )
-        ) : (
-          <>
-            {task ? (
-              <Button
-                icon={
-                  <span className={spacing.mrXs}>
-                    <ExclamationCircleIcon
-                      color={COLOR_HEX_VALUES_BY_NAME.blue}
-                    ></ExclamationCircleIcon>
-                  </span>
-                }
-                type="button"
-                variant="link"
-                onClick={navigateToAnalysisDetails}
-                className={spacing.ml_0}
-                style={{ margin: "0", padding: "0" }}
-              >
-                Analysis details
-              </Button>
-            ) : (
-              notAvailable
-            )}
-          </>
-        )}
-      </TextContent>
-
-      {!isFetching && !!facts.length && <ApplicationFacts facts={facts} />}
-    </>
   );
 };
